Moved methods.
